TuxOnIce: Fix swap allocator cleaning up.
When allocating swap, we don't add to the extent chains that record swap
entries until we come to a swap entry that requires us to start a new
extent (!= previous maximum + 1). This allows us to efficiently handle
stripped swap.

After allocating swap, we have a loop to go through, adding any extents
that were still 'in progress' (There'll always be one per swap device
that we allocated from).

We were breaking out of the loop after the adding the first extent. That
isn't right if we have more than one swap device in use. We ought to
handle them all.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>