TuxOnIce: Rename _toi_try_resume to toi_sys_power_disk_try_resume
In a previous patch, we renamed the function that software_resume itself
calls to get TuxOnIce to check whether to resume (the function that's
built in even if TuxOnIce is modular). Now rename the function in the
core module that gets invoked - from _toi_try_resume to
toi_sys_power_disk_try_resume, making it clear when this function is
used.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>