TuxOnIce: Don't BUG() if no more data to read.
In toi_bio_get_next_page_read, we were BUG()ing if there was no
page to read. In trying to find the cause of us hitting this BUG(),
I switched to returning -ENODATA, in the hope of getting a better
trace. For one user, at least, this (combined with the following
patches) has made the problem go away. I still need to get to the
root cause, but will commit these patches for now.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>