Add setup option to switch warning default.
One user has an issue where scsi drives will randomly be assigned their
major and minor numbers at boot time - sometimes they're swapped, sometimes
not. This causes TuxOnIce to be unable to find the image header, and the
"Failed to translate /dev/xxxx into a device id" message. Add a boot-time
option so that the default answer to this message is rebooting instead of
invalidating the image, so that he can easily retry until devices are
correctly detected.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>