Rework swap allocation for changes in vanilla.
The swap accounting changes in vanilla require changes in our
allocation routine. In particular, we need to use get_swap_page_of_type,
or our swap_free calls are ineffective. We therefore rework our swap
allocation algorithm so that we loop through available and usable swap
types as we seek to allocate the number of required pages, moving on to
the next type when all swap in one source of pages are allocated. Because
we no longer need to deal with striped swap, we no longer need the arrays
and no longer need the final loop.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>