No negative numbers in returning swap allocated.
The function for returning the amount of swap allocated could
potentially return a negative number if the amount of swap
available is less than our header reservation. In such cases,
return zero instead. This fixes the problem of the negative
numbers being converted to large positives and us thinking
we have enough swap when we don't.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>