Swap priority support.
Add support for utilising swap priorities in TuxOnIce. We remove the simple
striping that was added as a stop-gap, and add code that remembers and uses
priorities given at swapon time.

I plan to build on this in later patches, shifting a lot of the code that's
currently duplicated in the swap and file allocators to the block I/O module.
As part of this TuxOnIce will gain support for using both swap and a file
in the same image (you'll be able to specify a priority for the file, too).

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>