Fix resume failure iff first extent one page only
Fix a failure to resume that only occurs if the first extent on the
first chain contains only one page. In this case, we need to read
the second extent before we can set up the chain iteration state.
If we don't do that, we get into the eof state (because we skip the
first page, which was already read). This is almost certainly
Kenneth's transient resume failure.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>