Fix handling file allocator invalid targets.
Fix modifying the file allocator target, so that setting it to an
empty string or silly things like /dev/strerr doesn't cause an oops.
If you're silly enough to try /dev/stderr, you'll hang indefinitely
when you try to hibernate, but you won't cause an oops anymore by
simply setting the target file to that value.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>