LVM fix for new auto resume= support.
Francios reports a problem with LVM and the new automatic resume=
support not playing nicely together. Instead of trying to get the
proper path, let's just use the major and minor numbers that we
know will work. The user can always reset the value later, if they
want something nicer to look at.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>