Fix two bio allocation issues.
First, fix an issue with the last patch to this routine, wherein we
need to call devices_of_same_priority() once, in order to avoid
invoking it on a NULL chain in the routine's main loop.

Second, we shouldn't assume that allocators will actually find some
storage to register. Prio_chain_head might still be NULL after
invoking toi_bio_register_storage, so check for that.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>