Fix display of compression stats.
While testing, I noticed that compression stats displayed were only
values for pageset2, not the whole image. The data wasn't being
stored in the image header, but not transferred to the resumed
kernel via the boot kernel data structure. This patch addresses
that oversight, by implementing a generic method for putting
information into the structure pre-atommic-restore and using that
and the existing post-restore hook to get the data from the boot
kernel to the resume kernel, which can then display stats for the
complete image.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>