Add retry macro to try_to_open_resume_device
Help code readability and simplify future changes by introducing a macro
for try_to_open_resume_device that takes a command which is expected to
set resume_dev_t. It invokes the command and then checks whether resume_dev_t
was set. Iff resume_dev_t is still null and we haven't already waited for the
async device probing to complete, we wait on async device probing and then
perform the command again.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>