Fix gfp_flags used in getting ps1 load addresses.
There was one problem with the previous patch to the code for
getting pageset 1 load addresses: we were always using __GFP_HIGHMEM
in allocations, with the result being that we were allocating too
many highmem pages on systems with them. Fix this by removing the
highmem flag for the last $low_needed allocations, thereby ensuring
that those ones get lowmem pages.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>