Init & cleanup compression per pageset.
Rather than initialising compression once at the start of
hibernating or resuming, and cleaning up once, do this at the
beginning and end of reading or writing a pageset. The main
reason for doing things less efficiently is because if the
compressor being used isn't the default (lzo), then prior to this
change, we try and init lzo anyway before reading the name of
the compressor we'll actually use from the image header. If lzo
isn't compiled in, we'll error out. With this change, we'll never
even look at the lzo compression algo.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>