Add support for checking compression.
Add a new /sys/power/tuxonice/compress/check sysfs entry which defaults to
zero. If set non-zero, when writing an image, we check that the results of
compression, verifying that the result can be decompressed back to the
original contents. This will be used to help diagnose the cause of the
problems with reading an image that are sometimes seen.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>