Re-add LZF support.
While we're seeking to find the cause of problems with reading the image,
reinstate the LZF cryptoapi support. It has a long history of reliability
that should help eliminate one potential cause of issues (though it shouldn't
be necessary when the previous patch - implementing support for checking
decompressability - is used).

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>