Ignore I/O errors in UUID checking.
Our UUID code has until now BUGed if I/O fails when (for example)
scanning for a resumable image. This means (on my computer at least)
that if you start the computer with Charlie and Lola's Favourite
and Best Music Record in the CD drive (great for your 2 yo, btw!),
we hit the oops every time.

This patch adds proper handling of I/O errors, so that Charlie and
Lola can happily sit in the drive and we'll still scan other devices
as per normal.

Really, the most bestest and excellent solution that should have been
done from the start. :)

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>