Don't restart unnecessarily.
If the number of extra pages needed when doing the atomic copy
is greater than that expected, look to see if we can just allocate
the additional pages needed and continue. Only restart if we can't,
or if it looks like we now need to free some memory to fit within
storage constraints.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>