Correct usm header storage functions.
The userspace storage manager needs some love. The functions
for storing configuration in the image header weren't taking
account of the NULL on the end, and didn't also store the
length of the usm's path. I know it's not being used at the
moment, but it should still be right!

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>