Modify tracking of i/o count.
Modify the tracking of the amount of I/O still to do a little - when
reading the image, start by assuming we'll use the page. Only if
we're going to ignore a page do we reincrement the count. Make sure
we do all manipulation of the count under the mutex so the check
for exiting the loop is safe.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>