Extra atomic-copy time content recalculation
When preparing to do the atomic copy, we might find that the extra
pages allowance the user gave wasn't sufficient. In such cases,
we look to see if we can continue anyway, and allocate extra memory
for the copy if we can. We should then also do a recalculation of
the image contents, in case the allocation of extra memory included
an implicit allocation of new slab pages, which should also be saved.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>