Fix progress update with flusher thread.
When using a flusher thread, the thread with thread_num == 0
will be used for flushing, so the progress won't get updated
while I/O is being done. Fix this by counting the number of
threads that are actually submitting the I/O, and get the
first one in to do the updating of progress.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>