Remove old ZONE_START macro
It seems that this macro was used to fix non-x86 cases that
zone_start_pfn includes a PHYS_PFN_OFFSET, when the old suspend2 code
assumed that pfn was started from zero.

This is unnecessary now. In recent efforts that are trying to port
tuxonice to ARM OMAP4 architecture we have to redefine the ZONE_START
macro to not to subtract PHYS_PFN_OFFSET. I think it can be completely
removed now.

Signed-off-by: Kan-Ru Chen <kanru@0xlab.org>
Acked-by: Frank Hofmann <frank.hofmann@tomtom.com>
Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>