Handle unsorted highmem struct pages.
Changes upstream mean that highmem pfns do not now
necessarily get struct pages that are in increasing
order - the address of the struct page for pfn(n+1)
may be numerically less than the address of the struct
page for pfn(n). Up until now, we've been using the
struct page address in the orig_address of pbes and
relying on their ordering in searching for the
correct pbe. Switch to saving pfns instead, but only
for highmem pages (at least until I'm told the same
thing can happen for lowmem).

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>