Reduce diff to vanilla to just symbol exports.
kernel/power/process.c was at once stage quite different to
vanilla. Nowadays, all of the improvements have been merged
and the only changes we need to make are exporting symbols
for when TuxOnIce is built as modules. Fix a couple of mistakes
in recent merging of changes so that exports are indeed the only
modifications we make.

Signed-off-by: Nigel <nigel@tuxonice.net>