Further work on incremental images.
I realised as I started to write the code that the idea of calculating a quick
hash first as a way of avoiding a more expensive but more reliable hash isn't
going to work: if the quick hashes are the same, we need to calculate the
expensive hash to check the page has really not changed. If the quick hashes
are different, we need to calculate the expensive hash to store the new value
for next time round. So then, just calculate the SHA1 (or whatever) and just
use it.

Aside from that, this version does some more cleanup (the result of using the
compression support c file for a template) and adds some pseudo code for the
writing the image function.

Signed-off-by: Nigel <nigel@tuxonice.net>