Make the cyc2ns data pages be untracked by TOI.
The last piece of the puzzle - at least on my laptop. Loading microcode
can cause a printk, which can cause an attempt to get a timestamp for
the printk, which can make us try to write to per CPU cyc2ns data very
early in the startup of a secondary CPU. This can cause a triple fault.
I therefore mark that struct as untracked.

Signed-off-by: Nigel Cunningham <nigel@nigelcunningham.com.au>