zen-tune: Set minimum up_threshold to 10 for ondemand
For most systems, idle processor usage usually sits just under 10% per
core.  To reduce some of the ondemand frequency shuffling and to let the
processor run at its minimum voltage when mostly idle, set the minimum
up_threshold to 10.

This number is also up for changing.  If we get any reports for systems
behaving worse, we can drop this to 5 or 6.